+++ src/device/tsp/Tsp/Script/ScriptCompressor.cs
@@ -1,4 +1,4 @@
-using System.Text;
+using cc.isr.Std.EqualityExtensions;
 
 namespace cc.isr.VI.Tsp.Script;
 
@@ -24,7 +24,7 @@ public static class ScriptCompressor
     {
         if ( !string.IsNullOrWhiteSpace( prefix ) || !string.IsNullOrWhiteSpace( prefix ) )
         {
-            StringBuilder builder = new( prefix );
+            System.Text.StringBuilder builder = new( prefix );
             _ = builder.Append( compressedContents );
             _ = builder.Append( suffix );
             return builder.ToString();
@@ -54,7 +54,7 @@ public static class ScriptCompressor
         contents = cc.isr.Std.IO.Compression.StringCompressor.CompressToBase64( contents );
         if ( !string.IsNullOrWhiteSpace( prefix ) || !string.IsNullOrWhiteSpace( prefix ) )
         {
-            StringBuilder builder = new( prefix );
+            System.Text.StringBuilder builder = new( prefix );
             _ = builder.Append( contents );
             _ = builder.Append( suffix );
             return builder.ToString();
@@ -124,68 +124,6 @@ public static class ScriptCompressor
         return ScriptCompressor.Decompress( contents, ScriptCompressor.CompressedPrefix, ScriptCompressor.CompressedSuffix );
     }
 
-    /// <summary>   Compares to string using string readers. </summary>
-    /// <remarks>   2025-05-03. </remarks>
-    /// <param name="str1"> The first string. </param>
-    /// <param name="str2"> The second string. </param>
-    /// <returns>   True if it succeeds, false if it fails. </returns>
-    public static bool AreEqual( string str1, string str2 )
-    {
-        if ( str1 == null && str2 == null )
-            return true;
-        if ( str1 == null || str2 == null )
-            return false;
-        return ScriptCompressor.AreEqual( new StringReader( str1 ), new StringReader( str2 ) );
-    }
-
-    /// <summary>   Compares two strings using text readers. </summary>
-    /// <remarks>   2025-05-03. </remarks>
-    /// <param name="reader1">  The first reader. </param>
-    /// <param name="reader2">  The second reader. </param>
-    /// <returns>   True if it succeeds, false if it fails. </returns>
-    public static bool AreEqual( TextReader reader1, TextReader reader2 )
-    {
-        if ( reader1 == null && reader2 == null )
-            return true;
-        if ( reader1 == null || reader2 == null )
-            return false;
-        string line1;
-        string line2;
-        while ( reader1.Peek() != -1 && reader2.Peek() != -1 )
-        {
-            line1 = reader1.ReadLine();
-            line2 = reader2.ReadLine();
-            if ( line1 == null && line2 == null )
-                return true;
-            else if ( line1 == null || line2 == null )
-                return false;
-            else if ( !string.Equals( line1, line2, StringComparison.Ordinal ) )
-                return false;
-        }
-        return true;
-    }
-
-    /// <summary>   Determine if we are compressed equal. </summary>
-    /// <remarks>   2025-05-03. </remarks>
-    /// <param name="compressed1">          The first compressed. </param>
-    /// <param name="compressed2">          The second compressed. </param>
-    /// <returns>   True if compressed equal, false if not. </returns>
-    public static bool CompressedEqual( string compressed1, string compressed2 )
-    {
-        if ( compressed1 == null && compressed2 == null )
-            return true;
-        if ( compressed1 == null || compressed2 == null )
-            return false;
-        string decompressed1 = ScriptCompressor.Decompress( compressed1 );
-        string decompressed2 = ScriptCompressor.Decompress( compressed2 );
-        return ScriptCompressor.AreEqual( decompressed1, decompressed2 );
-
-        // string comparison failed when reader comparison did not fail!
-        // comparing the two strings in notepad++ we get equality.
-        // in a byte code file the last terminated line is ))()
-        // return string.Equals( decompressed1, decompressed2, StringComparison.Ordinal );
-    }
-
     /// <summary>   Compressed files equal. </summary>
     /// <remarks>   2025-05-03. </remarks>
     /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
@@ -204,7 +142,14 @@ public static class ScriptCompressor
             return true;
         else if ( System.IO.File.Exists( filePath1 ) && System.IO.File.Exists( filePath2 ) )
         {
-            return ScriptCompressor.CompressedEqual( System.IO.File.ReadAllText( filePath1 ), System.IO.File.ReadAllText( filePath2 ) );
+            return ScriptCompressor.Decompress( System.IO.File.ReadAllText( filePath1 ) ).LinesEqual(
+                ScriptCompressor.Decompress( System.IO.File.ReadAllText( filePath2 ) ) );
+
+            // we resorted to line-by-line comparison failed because the simple string comparison failed
+            // even though notepad++ showed the two strings of byte code as equal.
+            // string decompressed1 = ScriptCompressor.Decompress( System.IO.File.ReadAllText( filePath1 ) );
+            // string decompressed2 = ScriptCompressor.Decompress( System.IO.File.ReadAllText( filePath2 ) );
+            // return string.Equals( decompressed1, decompressed2, StringComparison.Ordinal );
         }
         else
             throw new FileNotFoundException( "One or both of the files do not exist.", filePath1 + " or " + filePath2 );