+++ src/vi/k2600/k2600.mstest/Tsp/TspSessionScriptTests.cs
@@ -146,7 +146,7 @@ public class TspSessionScriptTests : Device.Tests.Base.ScriptTests
 
             toFilePath = Path.Combine( folderPath, $"{toFilePath}c" );
             TestBase.ConsoleOutputMemberMessage( $"Compressing '{filePath}'\r\n\t\tto '{toFilePath}'" );
-            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.UTF8 );
+            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.Default );
 
             TestBase.ConsoleOutputMemberMessage( $"Importing script from compressed trimmed '{toFilePath}' file" );
             this.Device.Session.ImportScript( scriptName, toFilePath, TimeSpan.Zero );
@@ -194,7 +194,7 @@ public class TspSessionScriptTests : Device.Tests.Base.ScriptTests
 
             toFilePath = Path.Combine( folderPath, $"{filePath}c" );
             TestBase.ConsoleOutputMemberMessage( $"Compressing '{filePath}'\r\n\t\tto '{toFilePath}'" );
-            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.UTF8 );
+            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.Default );
 
         }
         catch
@@ -307,7 +307,7 @@ public class TspSessionScriptTests : Device.Tests.Base.ScriptTests
 
             toFilePath = Path.Combine( folderPath, $"{filePath}c" );
             TestBase.ConsoleOutputMemberMessage( $"Compressing '{filePath}'\r\n\t\tto '{toFilePath}'" );
-            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.UTF8 );
+            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.Default );
         }
         catch
         {
@@ -392,7 +392,7 @@ public class TspSessionScriptTests : Device.Tests.Base.ScriptTests
 
             string toFilePath = Path.Combine( folderPath, $"{filePath}c" );
             TestBase.ConsoleOutputMemberMessage( $"Compressing '{filePath}'\r\n\t\tto '{toFilePath}'" );
-            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.UTF8 );
+            System.IO.File.WriteAllText( toFilePath, ScriptCompressor.Compress( System.IO.File.ReadAllText( filePath ) ), System.Text.Encoding.Default );
 
             TestBase.ConsoleOutputMemberMessage( $"Importing script from compressed binary '{filePath}' file" );
             this.Device.Session.ImportScript( scriptName, toFilePath, TimeSpan.Zero );